Windows/requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
elevate==0.1.3
fastapi==0.63.0
fastapi-utils==0.2.1
h11==0.12.0
httplib2==0.18.1
idna==2.10
multidict==5.1.0
protobuf==3.14.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydantic==1.7.3
PyMsgBox==1.0.9
pyngrok==5.0.1
pyperclip==1.8.1
python-dotenv==0.15.0
python-oauth2==1.1.1
pytz==2020.5
pywin32==300
PyYAML==5.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
uvicorn==0.13.3
WMI==1.5.1
yarl==1.6.3
